ternary checks for null objects
if an object is decided to be non-fatal, it won't halt operation
    of the program. but we still need some check to make sure
    that we avoid null-pointer exceptions.

the objects that are known to be non-fatal:
    - OCR
    - USB
    - LED

actually, it's not the LED object, it's the SPI object inside.

actually, it's none of the above. the objects will always be
    created because you can't return NULL from a constructor.
    therefore, the only time that these are NULL is if we run
    out of memory.

I think, in retrospect, this is a really bad idea. but it may
    sow the seeds of a better implementation later on.